Matomo Analytics Frontend migration (#8566)
* autoregister plugin API


* add changeset


* use new frontend


* address copilot comments


* remove unused dependencies


* Update workspaces/analytics/plugins/analytics-module-matomo/src/plugin.test.ts


* Update workspaces/analytics/plugins/analytics-module-matomo/src/plugin.ts


* fix tests


* regen knip


---------
