+++ workspaces/analytics/.changeset/fix-matomo-api-registration.md
@@ -1,0 +1,5 @@
+---
+'@backstage-community/plugin-analytics-module-matomo': patch
+---
+
+Register the Matomo analytics API factory in the plugin object so it auto-registers when the plugin loads
