+++ workspaces/analytics/plugins/analytics-module-matomo/knip-report.md
@@ -1,8 +1,2 @@
 # Knip report
 
-## Unused devDependencies (1)
-
-| Name                    | Location          | Severity |
-| :---------------------- | :---------------- | :------- |
-| @backstage/core-app-api | package.json:59:6 | error    |
-
