+++ workspaces/analytics/plugins/analytics-module-matomo/package.json
@@ -23,9 +23,12 @@
     }
   },
   "backstage": {
-    "role": "frontend-plugin-module",
-    "pluginId": "app",
-    "pluginPackage": "@backstage/plugin-app"
+    "role": "frontend-plugin",
+    "pluginId": "analytics-module-matomo",
+    "pluginPackage": "@backstage-community/plugin-analytics-module-matomo",
+    "pluginPackages": [
+      "@backstage-community/plugin-analytics-module-matomo"
+    ]
   },
   "sideEffects": false,
   "scripts": {
@@ -56,7 +59,6 @@
   },
   "devDependencies": {
     "@backstage/cli": "backstage:^",
-    "@backstage/core-app-api": "backstage:^",
     "@backstage/core-components": "backstage:^",
     "@backstage/dev-utils": "backstage:^",
     "@testing-library/jest-dom": "6.9.1",
