--- workspaces/analytics/plugins/analytics-module-matomo/src/module.ts
@@ -1,41 +1,0 @@
-/*
- * Copyright 2025 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import {
-  configApiRef,
-  createFrontendModule,
-  identityApiRef,
-} from '@backstage/frontend-plugin-api';
-import { AnalyticsImplementationBlueprint } from '@backstage/plugin-app-react';
-
-import { MatomoAnalytics } from './api';
-
-const matomoImplementation = AnalyticsImplementationBlueprint.make({
-  name: 'matomo',
-  params: defineParams =>
-    defineParams({
-      deps: { configApi: configApiRef, identityApi: identityApiRef },
-      factory: ({ configApi, identityApi }) =>
-        MatomoAnalytics.fromConfig(configApi, { identityApi }),
-    }),
-});
-
-/**
- * @public
- */
-export const matomoModule = createFrontendModule({
-  pluginId: 'app',
-  extensions: [matomoImplementation],
-});
