+++ workspaces/analytics/plugins/analytics-module-matomo/src/plugin.test.ts
@@ -13,10 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { analyticsModuleMatomoPlugin } from './plugin';
+import {
+  analyticsModuleMatomoPlugin,
+  analyticsProviderMatomoPlugin,
+  MatomoAnalyticsApi,
+} from './plugin';
 
 describe('matomo', () => {
   it('should export plugin', () => {
     expect(analyticsModuleMatomoPlugin).toBeDefined();
+  });
+
+  it('should register the analytics API factory on the plugin', () => {
+    const apis = Array.from(analyticsModuleMatomoPlugin.getApis());
+    expect(apis).toContain(MatomoAnalyticsApi);
+  });
+
+  it('should export NFS plugin', () => {
+    expect(analyticsProviderMatomoPlugin).toBeDefined();
   });
 });
